Skip files which already exist on install

+++ install.py
@@ -8,6 +8,11 @@ targetdir = os.path.expanduser('~')
 for dotfile in os.listdir(sourcedir):
   source = os.path.join(sourcedir, dotfile)
   target = os.path.join(targetdir, dotfile)
+
+  if os.path.exists(target):
+      print "WARNING: %s exists! Skipping." % (target)
+      continue
+
   ##TODO## :: Look at why os.symlink doesn't like relative paths
   os.system("ln -s %s %s" % (source, target))
